Update movein_fslinks.sh source dir
- change source directory from /home/data_ext/home_russell_backups to
/home/data_ext/home/russell/

- Add some basic links for vim if Dropbox directory exists, but fslinks.txt
doesn't

+++ movein_fslinks.sh
@@ -1,21 +1,21 @@
 #!/bin/bash
 
-# This is a script to assist with setting up a new Linux computer, at FCI. It 
-# attempts to install a bunch of packages, setup networking, setup ssh, and
-# copy some config files from a base user's home directory on a remote 
-# host/server.
+# This is a script to assist with setting up a new Linux computer by setting up
+# some symbolic links
 
 # !!!!!!!!!!!!!!!!! THIS SCRIPT IS NOT TESTED !!!!!!!!!!!!!!!!!!
 # This is just a starter, for next time. - RBB 2013-10-11
 
 
 
-if [ -d "/home/data_ext/home_russell_backup" ]; then
+if [ -d "/home/data_ext/home/russell" ]; then
    if [ ! -f "$HOME/fslinks.txt" ]; then
-      cp -ap /home/data_ext/home_russell_backup/fslinks.txt "$HOME" 
+      echo "Copying fslinks.txt"
+      cp -ap /home/data_ext/home/russell/fslinks.txt "$HOME" 
    fi
    if [ ! -f "$HOME/installed.txt" ]; then
-      cp -ap /home/data_ext/home_russell_backup/installed.txt "$HOME" 
+      echo "Copying installed.txt"
+      cp -ap /home/data_ext/home/russell/installed.txt "$HOME" 
    fi
 fi
 
@@ -55,28 +55,48 @@ mvl_home_target () {
 
 #Created by backext.sh. This is a list of symbolic links in the src directory
 #lrwxrwxrwx 1 russell russell 15 Sep 18  2017 ./.bashrc -> Dropbox/.bashrc*
-while read line; do
-   #echo $line
-   src=$(echo $line | grep -v "Created by backext.sh" | cut -d ' ' -f 11)
-   dst=$(echo $line | grep -v "Created by backext.sh" | cut -d ' ' -f 9)
+if [ -e "$HOME/fslinks.txt" ]; then
+   while read line; do
+      #echo $line
+      src=$(echo $line | grep -v "Created by backext.sh" | cut -d ' ' -f 11)
+      dst=$(echo $line | grep -v "Created by backext.sh" | cut -d ' ' -f 9)
 
-   if [ ! -z $src ]; then
-      base=$(basename $dst)
-      src=$(realpath -s $src)
-      dst=$(realpath -s $dst)
-      #echo "src = $src"
-      #echo "dst = $dst"
-      #echo "base = $base"
+      if [ ! -z $src ]; then
+         base=$(basename $dst)
+         src=$(realpath -s $src)
+         dst=$(realpath -s $dst)
+         #echo "src = $src"
+         #echo "dst = $dst"
+         #echo "base = $base"
 
-      mv_home_target "$base"
+         mv_home_target "$base"
 
-      #echo "src and dst not empty"
-      if [ ! -e "$dst" ]; then
-         echo "Creating link $src -> $dst"
-         ln -s "$src" "$dst"
-      else
-         echo "$dst exists"
+         #echo "src and dst not empty"
+         if [ ! -e "$dst" ]; then
+            echo "Creating link $src -> $dst"
+            ln -s "$src" "$dst"
+         else
+            echo "$dst exists"
+         fi
       fi
+      echo "------"
+   done < "$HOME/fslinks.txt"
+else
+   # Vim setup
+   if [ -d "$HOME/Dropbox" ]; then
+      if [ -e "$HOME/.vimrc" ]; then
+         mv "$HOME/.vimrc" "$HOME/.vimrc_movein_fslinks.bak"
+      fi
+      ln -s "$HOME/Dropbox/_vimrc" "$HOME/.vimrc"
+
+      if [ -e "$HOME/.gvimrc" ]; then
+         mv "$HOME/.gvimrc" "$HOME/.gvimrc_movein_fslinks.bak"
+      fi
+      ln -s "$HOME/Dropbox/_gvimrc" "$HOME/.gvimrc"
+
+      if [ -e "$HOME/.vim" ]; then
+         mv "$HOME/.vim" "$HOME/.vim_movein_fslinks.bak"
+      fi
+      ln -s "$HOME/Dropbox/vimfiles" "$HOME/.vim"
    fi
-   echo "------"
-done < "$HOME/fslinks.txt"
+fi
